// Ref/Web/ReferenceDataUpdateService.Web/src/models/RefCusProcedureAttributeUserView.ts
//------------------------------------------------------------------------------
// <auto-generated>
//     This code was generated from a template.
//
//     Manual changes to this file may cause unexpected behavior in your application.
//     Manual changes to this file will be overwritten if the code is regenerated.
// </auto-generated>
//------------------------------------------------------------------------------
import { IEntity } from "./IEntity";

export class RefCusProcedureAttributeUserView implements IEntity {
	ZXB_PK: string;
	ZXB_ZZ6_ProcedureCode: string;
	ZXB_Name: string;
	static readonly ZXB_Name_MaxLength: number = 50;
	ZXB_Value: string;
	ZXB_CountryOrGrouping: string;
	static readonly ZXB_CountryOrGrouping_MaxLength: number = 3;
	ZXB_IsEditable: boolean;
}