// Ref/Web/ReferenceDataUpdateService.Web/src/models/RefCusProcedureUserView.ts
//------------------------------------------------------------------------------
// <auto-generated>
//     This code was generated from a template.
//
//     Manual changes to this file may cause unexpected behavior in your application.
//     Manual changes to this file will be overwritten if the code is regenerated.
// </auto-generated>
//------------------------------------------------------------------------------
import { IEntity } from "./IEntity";

export class RefCusProcedureUserView implements IEntity {
	ZZ6_PK: string;
	ZZ6_Category: string;
	static readonly ZZ6_Category_MaxLength: number = 3;
	ZZ6_ProcedureCode: string;
	static readonly ZZ6_ProcedureCode_MaxLength: number = 5;
	ZZ6_PreviousProcedureCode: string;
	static readonly ZZ6_PreviousProcedureCode_MaxLength: number = 5;
	ZZ6_Concession: string;
	static readonly ZZ6_Concession_MaxLength: number = 50;
	ZZ6_Description: string;
	ZZ6_CountryOrGrouping: string;
	static readonly ZZ6_CountryOrGrouping_MaxLength: number = 3;
	ZZ6_ShipmentType: string;
	static readonly ZZ6_ShipmentType_MaxLength: number = 50;
	ZZ6_CalculateDuty: boolean;
	ZZ6_Group: string;
	static readonly ZZ6_Group_MaxLength: number = 200;
	ZZ6_LandedCost: boolean;
	ZZ6_IntoWarehouse: string;
	static readonly ZZ6_IntoWarehouse_MaxLength: number = 1;
	ZZ6_OutOfWarehouse: string;
	static readonly ZZ6_OutOfWarehouse_MaxLength: number = 1;
	ZZ6_IntoInwardProcessing: string;
	static readonly ZZ6_IntoInwardProcessing_MaxLength: number = 1;
	ZZ6_OutOfInwardProcessing: string;
	static readonly ZZ6_OutOfInwardProcessing_MaxLength: number = 1;
	ZZ6_IntoOutwardProcessing: string;
	static readonly ZZ6_IntoOutwardProcessing_MaxLength: number = 1;
	ZZ6_OutofOutwardProcessing: string;
	static readonly ZZ6_OutofOutwardProcessing_MaxLength: number = 1;
	ZZ6_IntoTemporaryImport: string;
	static readonly ZZ6_IntoTemporaryImport_MaxLength: number = 1;
	ZZ6_OutOfTemporaryImport: string;
	static readonly ZZ6_OutOfTemporaryImport_MaxLength: number = 1;
	ZZ6_IntoTemporaryExport: string;
	static readonly ZZ6_IntoTemporaryExport_MaxLength: number = 1;
	ZZ6_OutOfTemporaryExport: string;
	static readonly ZZ6_OutOfTemporaryExport_MaxLength: number = 1;
	ZZ6_StartDate: string;
	ZZ6_EndDate: string;
	ZZ6_CalculateVAT: boolean;
	ZZ6_IsGuaranteeConsumed: string;
	static readonly ZZ6_IsGuaranteeConsumed_MaxLength: number = 1;
	ZZ6_IsGuaranteeReleased: string;
	static readonly ZZ6_IsGuaranteeReleased_MaxLength: number = 1;
	ZZ6_IsTransit: string;
	static readonly ZZ6_IsTransit_MaxLength: number = 1;
	ZZ6_IsPublished: boolean;
	ZZ6_IsEditable: boolean;
}